fix the unzip in one of the functions
this future proof the unzip I had where I only designated one variable. Since two are there, need two variables. Python 2 didn't trigger an error but with Python 3 it did. I don't know if for Python 2, it actually only took the first variable as I was trying to do or it just never triggered an error.